[rename] #48 User 인증, 검증 컨트롤러로 기존 컨트롤러 이름 변경
